[net] Move SOCKS5 proxy code into socks5.{h,cpp}
