+++ src/netbase.cpp
@@ -30,15 +30,8 @@
 #endif
 
 // Settings
-static GlobalMutex g_proxyinfo_mutex;
-static Proxy proxyInfo[NET_MAX] GUARDED_BY(g_proxyinfo_mutex);
-static Proxy nameProxy GUARDED_BY(g_proxyinfo_mutex);
 int nConnectTimeout = DEFAULT_CONNECT_TIMEOUT;
 bool fNameLookup = DEFAULT_NAME_LOOKUP;
-
-// Need ample time for negotiation for very slow proxies such as Tor (milliseconds)
-int g_socks5_recv_timeout = 20 * 1000;
-static std::atomic<bool> interruptSocks5Recv(false);
 
 std::vector<CNetAddr> WrappedGetAddrInfo(const std::string& name, bool allow_lookup)
 {
@@ -240,248 +233,6 @@ CService LookupNumeric(const std::string& name, uint16_t portDefault, DNSLookupF
     return addr;
 }
 
-/** SOCKS version */
-enum SOCKSVersion: uint8_t {
-    SOCKS4 = 0x04,
-    SOCKS5 = 0x05
-};
-
-/** Values defined for METHOD in RFC1928 */
-enum SOCKS5Method: uint8_t {
-    NOAUTH = 0x00,        //!< No authentication required
-    GSSAPI = 0x01,        //!< GSSAPI
-    USER_PASS = 0x02,     //!< Username/password
-    NO_ACCEPTABLE = 0xff, //!< No acceptable methods
-};
-
-/** Values defined for CMD in RFC1928 */
-enum SOCKS5Command: uint8_t {
-    CONNECT = 0x01,
-    BIND = 0x02,
-    UDP_ASSOCIATE = 0x03
-};
-
-/** Values defined for REP in RFC1928 */
-enum SOCKS5Reply: uint8_t {
-    SUCCEEDED = 0x00,        //!< Succeeded
-    GENFAILURE = 0x01,       //!< General failure
-    NOTALLOWED = 0x02,       //!< Connection not allowed by ruleset
-    NETUNREACHABLE = 0x03,   //!< Network unreachable
-    HOSTUNREACHABLE = 0x04,  //!< Network unreachable
-    CONNREFUSED = 0x05,      //!< Connection refused
-    TTLEXPIRED = 0x06,       //!< TTL expired
-    CMDUNSUPPORTED = 0x07,   //!< Command not supported
-    ATYPEUNSUPPORTED = 0x08, //!< Address type not supported
-};
-
-/** Values defined for ATYPE in RFC1928 */
-enum SOCKS5Atyp: uint8_t {
-    IPV4 = 0x01,
-    DOMAINNAME = 0x03,
-    IPV6 = 0x04,
-};
-
-/** Status codes that can be returned by InterruptibleRecv */
-enum class IntrRecvError {
-    OK,
-    Timeout,
-    Disconnected,
-    NetworkError,
-    Interrupted
-};
-
-/**
- * Try to read a specified number of bytes from a socket. Please read the "see
- * also" section for more detail.
- *
- * @param data The buffer where the read bytes should be stored.
- * @param len The number of bytes to read into the specified buffer.
- * @param timeout The total timeout in milliseconds for this read.
- * @param sock The socket (has to be in non-blocking mode) from which to read bytes.
- *
- * @returns An IntrRecvError indicating the resulting status of this read.
- *          IntrRecvError::OK only if all of the specified number of bytes were
- *          read.
- *
- * @see This function can be interrupted by calling InterruptSocks5(bool).
- *      Sockets can be made non-blocking with Sock::SetNonBlocking().
- */
-static IntrRecvError InterruptibleRecv(uint8_t* data, size_t len, int timeout, const Sock& sock)
-{
-    int64_t curTime = GetTimeMillis();
-    int64_t endTime = curTime + timeout;
-    while (len > 0 && curTime < endTime) {
-        ssize_t ret = sock.Recv(data, len, 0); // Optimistically try the recv first
-        if (ret > 0) {
-            len -= ret;
-            data += ret;
-        } else if (ret == 0) { // Unexpected disconnection
-            return IntrRecvError::Disconnected;
-        } else { // Other error or blocking
-            int nErr = WSAGetLastError();
-            if (nErr == WSAEINPROGRESS || nErr == WSAEWOULDBLOCK || nErr == WSAEINVAL) {
-                // Only wait at most MAX_WAIT_FOR_IO at a time, unless
-                // we're approaching the end of the specified total timeout
-                const auto remaining = std::chrono::milliseconds{endTime - curTime};
-                const auto timeout = std::min(remaining, std::chrono::milliseconds{MAX_WAIT_FOR_IO});
-                if (!sock.Wait(timeout, Sock::RECV)) {
-                    return IntrRecvError::NetworkError;
-                }
-            } else {
-                return IntrRecvError::NetworkError;
-            }
-        }
-        if (interruptSocks5Recv)
-            return IntrRecvError::Interrupted;
-        curTime = GetTimeMillis();
-    }
-    return len == 0 ? IntrRecvError::OK : IntrRecvError::Timeout;
-}
-
-/** Convert SOCKS5 reply to an error message */
-static std::string Socks5ErrorString(uint8_t err)
-{
-    switch(err) {
-        case SOCKS5Reply::GENFAILURE:
-            return "general failure";
-        case SOCKS5Reply::NOTALLOWED:
-            return "connection not allowed";
-        case SOCKS5Reply::NETUNREACHABLE:
-            return "network unreachable";
-        case SOCKS5Reply::HOSTUNREACHABLE:
-            return "host unreachable";
-        case SOCKS5Reply::CONNREFUSED:
-            return "connection refused";
-        case SOCKS5Reply::TTLEXPIRED:
-            return "TTL expired";
-        case SOCKS5Reply::CMDUNSUPPORTED:
-            return "protocol error";
-        case SOCKS5Reply::ATYPEUNSUPPORTED:
-            return "address type not supported";
-        default:
-            return "unknown";
-    }
-}
-
-bool Socks5(const std::string& strDest, uint16_t port, const ProxyCredentials* auth, const Sock& sock)
-{
-    IntrRecvError recvr;
-    LogPrint(BCLog::NET, "SOCKS5 connecting %s\n", strDest);
-    if (strDest.size() > 255) {
-        return error("Hostname too long");
-    }
-    // Construct the version identifier/method selection message
-    std::vector<uint8_t> vSocks5Init;
-    vSocks5Init.push_back(SOCKSVersion::SOCKS5); // We want the SOCK5 protocol
-    if (auth) {
-        vSocks5Init.push_back(0x02); // 2 method identifiers follow...
-        vSocks5Init.push_back(SOCKS5Method::NOAUTH);
-        vSocks5Init.push_back(SOCKS5Method::USER_PASS);
-    } else {
-        vSocks5Init.push_back(0x01); // 1 method identifier follows...
-        vSocks5Init.push_back(SOCKS5Method::NOAUTH);
-    }
-    ssize_t ret = sock.Send(vSocks5Init.data(), vSocks5Init.size(), MSG_NOSIGNAL);
-    if (ret != (ssize_t)vSocks5Init.size()) {
-        return error("Error sending to proxy");
-    }
-    uint8_t pchRet1[2];
-    if (InterruptibleRecv(pchRet1, 2, g_socks5_recv_timeout, sock) != IntrRecvError::OK) {
-        LogPrintf("Socks5() connect to %s:%d failed: InterruptibleRecv() timeout or other failure\n", strDest, port);
-        return false;
-    }
-    if (pchRet1[0] != SOCKSVersion::SOCKS5) {
-        return error("Proxy failed to initialize");
-    }
-    if (pchRet1[1] == SOCKS5Method::USER_PASS && auth) {
-        // Perform username/password authentication (as described in RFC1929)
-        std::vector<uint8_t> vAuth;
-        vAuth.push_back(0x01); // Current (and only) version of user/pass subnegotiation
-        if (auth->username.size() > 255 || auth->password.size() > 255)
-            return error("Proxy username or password too long");
-        vAuth.push_back(auth->username.size());
-        vAuth.insert(vAuth.end(), auth->username.begin(), auth->username.end());
-        vAuth.push_back(auth->password.size());
-        vAuth.insert(vAuth.end(), auth->password.begin(), auth->password.end());
-        ret = sock.Send(vAuth.data(), vAuth.size(), MSG_NOSIGNAL);
-        if (ret != (ssize_t)vAuth.size()) {
-            return error("Error sending authentication to proxy");
-        }
-        LogPrint(BCLog::PROXY, "SOCKS5 sending proxy authentication %s:%s\n", auth->username, auth->password);
-        uint8_t pchRetA[2];
-        if (InterruptibleRecv(pchRetA, 2, g_socks5_recv_timeout, sock) != IntrRecvError::OK) {
-            return error("Error reading proxy authentication response");
-        }
-        if (pchRetA[0] != 0x01 || pchRetA[1] != 0x00) {
-            return error("Proxy authentication unsuccessful");
-        }
-    } else if (pchRet1[1] == SOCKS5Method::NOAUTH) {
-        // Perform no authentication
-    } else {
-        return error("Proxy requested wrong authentication method %02x", pchRet1[1]);
-    }
-    std::vector<uint8_t> vSocks5;
-    vSocks5.push_back(SOCKSVersion::SOCKS5); // VER protocol version
-    vSocks5.push_back(SOCKS5Command::CONNECT); // CMD CONNECT
-    vSocks5.push_back(0x00); // RSV Reserved must be 0
-    vSocks5.push_back(SOCKS5Atyp::DOMAINNAME); // ATYP DOMAINNAME
-    vSocks5.push_back(strDest.size()); // Length<=255 is checked at beginning of function
-    vSocks5.insert(vSocks5.end(), strDest.begin(), strDest.end());
-    vSocks5.push_back((port >> 8) & 0xFF);
-    vSocks5.push_back((port >> 0) & 0xFF);
-    ret = sock.Send(vSocks5.data(), vSocks5.size(), MSG_NOSIGNAL);
-    if (ret != (ssize_t)vSocks5.size()) {
-        return error("Error sending to proxy");
-    }
-    uint8_t pchRet2[4];
-    if ((recvr = InterruptibleRecv(pchRet2, 4, g_socks5_recv_timeout, sock)) != IntrRecvError::OK) {
-        if (recvr == IntrRecvError::Timeout) {
-            /* If a timeout happens here, this effectively means we timed out while connecting
-             * to the remote node. This is very common for Tor, so do not print an
-             * error message. */
-            return false;
-        } else {
-            return error("Error while reading proxy response");
-        }
-    }
-    if (pchRet2[0] != SOCKSVersion::SOCKS5) {
-        return error("Proxy failed to accept request");
-    }
-    if (pchRet2[1] != SOCKS5Reply::SUCCEEDED) {
-        // Failures to connect to a peer that are not proxy errors
-        LogPrintf("Socks5() connect to %s:%d failed: %s\n", strDest, port, Socks5ErrorString(pchRet2[1]));
-        return false;
-    }
-    if (pchRet2[2] != 0x00) { // Reserved field must be 0
-        return error("Error: malformed proxy response");
-    }
-    uint8_t pchRet3[256];
-    switch (pchRet2[3])
-    {
-        case SOCKS5Atyp::IPV4: recvr = InterruptibleRecv(pchRet3, 4, g_socks5_recv_timeout, sock); break;
-        case SOCKS5Atyp::IPV6: recvr = InterruptibleRecv(pchRet3, 16, g_socks5_recv_timeout, sock); break;
-        case SOCKS5Atyp::DOMAINNAME:
-        {
-            recvr = InterruptibleRecv(pchRet3, 1, g_socks5_recv_timeout, sock);
-            if (recvr != IntrRecvError::OK) {
-                return error("Error reading from proxy");
-            }
-            int nRecv = pchRet3[0];
-            recvr = InterruptibleRecv(pchRet3, nRecv, g_socks5_recv_timeout, sock);
-            break;
-        }
-        default: return error("Error: malformed proxy response");
-    }
-    if (recvr != IntrRecvError::OK) {
-        return error("Error reading from proxy");
-    }
-    if (InterruptibleRecv(pchRet3, 2, g_socks5_recv_timeout, sock) != IntrRecvError::OK) {
-        return error("Error reading from proxy");
-    }
-    LogPrint(BCLog::NET, "SOCKS5 connected %s\n", strDest);
-    return true;
-}
-
 std::unique_ptr<Sock> CreateSockTCP(const CService& address_family)
 {
     // Create a sockaddr from the specified service.
@@ -610,76 +361,6 @@ bool ConnectSocketDirectly(const CService &addrConnect, const Sock& sock, int nT
     return true;
 }
 
-bool SetProxy(enum Network net, const Proxy &addrProxy) {
-    assert(net >= 0 && net < NET_MAX);
-    if (!addrProxy.IsValid())
-        return false;
-    LOCK(g_proxyinfo_mutex);
-    proxyInfo[net] = addrProxy;
-    return true;
-}
-
-bool GetProxy(enum Network net, Proxy &proxyInfoOut) {
-    assert(net >= 0 && net < NET_MAX);
-    LOCK(g_proxyinfo_mutex);
-    if (!proxyInfo[net].IsValid())
-        return false;
-    proxyInfoOut = proxyInfo[net];
-    return true;
-}
-
-bool SetNameProxy(const Proxy &addrProxy) {
-    if (!addrProxy.IsValid())
-        return false;
-    LOCK(g_proxyinfo_mutex);
-    nameProxy = addrProxy;
-    return true;
-}
-
-bool GetNameProxy(Proxy &nameProxyOut) {
-    LOCK(g_proxyinfo_mutex);
-    if(!nameProxy.IsValid())
-        return false;
-    nameProxyOut = nameProxy;
-    return true;
-}
-
-bool HaveNameProxy() {
-    LOCK(g_proxyinfo_mutex);
-    return nameProxy.IsValid();
-}
-
-bool IsProxy(const CNetAddr &addr) {
-    LOCK(g_proxyinfo_mutex);
-    for (int i = 0; i < NET_MAX; i++) {
-        if (addr == static_cast<CNetAddr>(proxyInfo[i].proxy))
-            return true;
-    }
-    return false;
-}
-
-bool ConnectThroughProxy(const Proxy& proxy, const std::string& strDest, uint16_t port, const Sock& sock, int nTimeout, bool& outProxyConnectionFailed)
-{
-    // first connect to proxy server
-    if (!ConnectSocketDirectly(proxy.proxy, sock, nTimeout, true)) {
-        outProxyConnectionFailed = true;
-        return false;
-    }
-    // do socks negotiation
-    if (proxy.randomize_credentials) {
-        ProxyCredentials random_auth;
-        static std::atomic_int counter(0);
-        random_auth.username = random_auth.password = strprintf("%i", counter++);
-        if (!Socks5(strDest, port, &random_auth, sock)) {
-            return false;
-        }
-    } else {
-        if (!Socks5(strDest, port, nullptr, sock)) {
-            return false;
-        }
-    }
-    return true;
-}
 
 bool LookupSubNet(const std::string& subnet_str, CSubNet& subnet_out)
 {
@@ -714,11 +395,6 @@ bool LookupSubNet(const std::string& subnet_str, CSubNet& subnet_out)
         }
     }
     return false;
-}
-
-void InterruptSocks5(bool interrupt)
-{
-    interruptSocks5Recv = interrupt;
 }
 
 bool IsBadPort(uint16_t port)
